% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Name: Kévin Allan Sales Rodrigues}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The filename of the .csv file you want to import}
}
\value{
This function returns an R tibble containing the data from the .csv
  file that was read in. This function will return Error in
  file.exists(filename) : invalid 'file' argument if no file exists at the
  filepath specified as the filename argument.
}
\description{
Name: Kévin Allan Sales Rodrigues
}
\examples{
\dontrun{fars_read("coursera/data/accident_2013.csv.bz2")}
\dontrun{fars_read("data.csv")}

}
