% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Import data for multiple years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{The list or vector of four-digit years for which you want to
import information.}
}
\value{
This function returns a list in which each element corresponds to
  an element in the argument "years." For each valid year in "years," the
  corresponding element of the list will be a tibble with rows equal to the
  number of observations and two columns: MONTH and year.

  If an invalid year
  is provided as an argument, the corresponding element of the returned list
  will be NULL. A warning message will then print: \code{Warning message: In
  value[[3L]](cond) : invalid year}
}
\description{
Import data for multiple years
}
\examples{
\dontrun{
fars_read_years(2015)
fars_read_years(c(2013, 2014))
fars_read_years(c(2013, "invalid input"))
}

}
