% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make Filename From Year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The year you want to pull data from.}
}
\value{
This function returns a valid filepath within
your R working directory.
}
\description{
This function takes a year as an argument and uses that to
construct a valid filepath to the accident dataset for that year.
}
\examples{
\dontrun{
make_filename(2015)
make_filename(2014)
make_filename(10)
}
}
